[cast] Properly cast to the correct return value

+++ common/core/syscall_model.cc
@@ -290,7 +290,7 @@ bool SyscallMdl::runEnter(IntPtr syscall_number, syscall_args_t &args)
          break;
       }
 
-      case -1:
+      case static_cast<unsigned long>(-1):
       default:
          break;
    }
